requirements.txt
flask
flask-socketio
requests
python-socketio[client]
redis